Return target_info from ntlm_decode_auth_msg
The target_info structure embedded in the NT Response message in NTLMv2
contains information needed to establish if the client has sent a valid
MIC. So we need to extract and return it if the caller requested it.

Also moves some wire structures definitions in common to be able to
reuse them.