Fix order of signature vs payload
The code was dead wrong and putting the cart before the horses.
The correct framing is to put the signature first an then the encrypted
payload. we were doing the opposite ... how embarrassing.

A milliong thanks to David Woodhouse for his persistence in testing and
assisting in finding out the issue.