Names with a . in the domain are enteprise names
This allows people to put in an email address as the source name and
have i treated automatically as an enterprise name as well.

Although technically NetBIOS names can have dots it is unlikely and the
user@domain form is generally undestood to be used with UPNs and email
like addresses which use the DNS Domain Name.

The fallback case for NetBIOS domain names with a dot is to configure the
client to use the DOMAIN\user name form instead.