Fix length check of nt_response
An array passed as a function argument is just a cosmetic ay to pass just a
pointer. Therefore sizeof(array) will only return the pointer length, not
the array length, and on 32 bit pointers are 4 bytes long.

Fix payload calculation by passing in the known correct length instead of using
fancy sizeofs ...