Drop support for GSS_C_MA_NOT_DFLT_MECH
When this MEchanism Attribute is exposed as supported it causes GSSAPI
implementations to eclude the mechanism completely as vaiable to acquire
creds from pseudo mechanisms like SPNEGO.
This is not what was intended when we added this flag. The intention was
to not make this mechanism the default when mutliple are available, and
specifically to not make it preferred over krb5.
However given this is not how the fal works, we need to drop it now.

It may be re-introduced at a later time as a runtime settings when we
grow a way to set configuration in a file somewhere so that admin can
control this behavior.

Signed-off-by: Simo Sorce <simo@redhat.com>