Change license to the more permissive ISC
This change is not sudden, I have been thinking about relicensing this
project for quite a while.
The main trigger is that I received a couple of requests for a more
permissive license, but the real reason is that I always felt it at
odds that a module that is targeted to be loaded by MIT/Krb5's gssapi
library had such a widely different license.
It may give rise to misunderstandings as well as create undue burden
on users to figure out license compatibility with their applications
(proven in fact by the requests to relicense).

Signed-off-by: Simo Sorce <simo@redhat.com>