Do not require cached users with winbind
When acquiring credentials we may need to know if a cache is available.
This is necessary in the ISC cacse as we need to use those cached
credentials to perform NTLM authentication. However in the ASC case
we do not need creds of the user to be cached, all we need is a handle
that can be used to later authenticated the user against a DC.

Signed-off-by: Simo Sorce <simo@redhat.com>