Adjust serialization for name attributes
This allows us to serialize and import back also named attributes.

NOTE: this commit also fixes a few severe export_data_buffer reallocation
bugs that so far flew under the radar because the initial allocation
size (4k) was big enough to contain the context as used in tests.
Larger contexts caused failures in the code that were also hard to
diagnose.

Signed-off-by: Simo Sorce <simo@redhat.com>