Use thread local storage for winbind context
This allow multiple calls into winbindd from parallel threads without
suffering from a common mutex that serializaes all authentications.

The price to pay is that a new context is allocated for each thread, and
the way windbind works, this probably mean a separate file descriptor is
created for each of this thread on the application side (winbindd will
be ok as it will kick idle connections).

The users of gssntlmssp in a multi-threaded environment will need to be
careful to limit the number of threads they allow to use gssapi in this
case.

Signed-off-by: Simo Sorce <simo@redhat.com>