Fix offset calculation for the MIC field to match the specification
In most cases the change is bening because NTLMSSP_NEGOTIATE_VERSION
is sent by the server in CHALLENGE_MESSAGE and then repeated in
AUTHENTICATION_MESSAGE which results in the correct offset being used.

To reduce the chance of breaking implementations that don't send the
NTLMSSP_NEGOTIATE_VERSION flag in CHALLENGE_MESSAGE it is also added
to NEGOTIATE_MESSAGE in the initial exchange.