Add option to set the default NEGOTIATE flags
This requires to explicitly acquire credentials before calling into
the ISC or ASC functions. Then setting the new default flags on the
credntial. It is not elegant to mix credentials and context flags,
but it is the only way to do it in GSSAPI currently, and there is
precedent in other mechanisms like KRB5.

Signed-off-by: Simo Sorce <simo@redhat.com>