Make gssntlm_import_name_by_mech more robust
This was reported by the GitHub Security Lab team as:
- Memory corruption when importing host-based service names (GHSL-2023-014)

However the only caller that could cause the issue report in a way that
culd lead to memory corruption is the application itself calling
gss_import_name with an invalid name buffer containing embedded zeros.

This is an invalid name for any GSSAPI mechanism and an incorrect use or
incorrect validation on the application side. As such I do not believe
this is a security or DoS issue in the gss-ntlmssp code, but will adopt
the proposed mitigation anyway as it makes the code more robust in the
face of application mistakes.

Signed-off-by: Simo Sorce <simo@redhat.com>